+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "PHP Sniffer" extension will be documented in this fi
 - Reword ENOENT errors
 - Add new setting `disableWhenDebugging` to disable `phpcs` when any debug session is active (#42)
 - Add option to disable validation (#38)
+- Add setting for running on non-PHP files (#16)
 
 ### Fixed
 - Avoid "write EPIPE" error (#35)